requirements.txt
openai==0.28.0
flask
python-dotenv
flask
supabase
pyairtable